significant changes to import script... call to it from appy.py in progress
