+++ notes
@@ -268,13 +268,10 @@ dbscan_stuff(coordinates)
 file uploads
 . files get uploaded to endpoint /api/users/<user>/albums/<album>/photos 
 . server:
-		TRY
 	-- save files to UPLOAD_DIR/TEMPDIR (python module tempdir for TEMPDIR)
-		TRY
 	-- process files -- use existing import_data code
 		-- put json in db
  		-- send to s3
- 		TRY
 	-- delete tempdir
